feat: 🎸 (Amount) 추가 나머지 금액 계산, 몫 계산, 0원 유무, 퍼센트 계산 메서드 작성
